Chang extention of React comonent files .js -> .jsx
