--- client/index.js
@@ -1,5 +1,0 @@
-import ReactDOM from 'react-dom';
-import React from 'react';
-import ItemView from './components/itemView.js'
-
-ReactDOM.render(<ItemView/>, document.getElementById('app'));
