+++ client/index.jsx
@@ -1,0 +1,5 @@
+import ReactDOM from 'react-dom';
+import React from 'react';
+import ItemView from './components/itemView.jsx';
+
+ReactDOM.render(<ItemView />, document.getElementById('app'));
