+++ webpack.config.js
@@ -3,7 +3,7 @@ const path = require('path');
 
 module.exports = {
   context: __dirname + '/client',
-  entry: './index.js',
+  entry: './index.jsx',
   module: {
     loaders: [
       {
